Remove abstract class MediumMultipleFormatter
