--- src/Multiple/MediumMultipleFormatter.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace CultuurNet\CalendarSummaryV3\Multiple;
-
-abstract class MediumMultipleFormatter
-{
-    /**
-     * @var string $langCode
-     */
-    protected $langCode;
-
-    /**
-     * @var bool $hidepast
-     */
-    protected $hidePast;
-
-    public function __construct(string $langCode, bool $hidePastDates)
-    {
-        $this->langCode = $langCode;
-        $this->hidePast = $hidePastDates;
-    }
-}
